+++ packages/espressocash_app/lib/features/accounts/services/accounts_event.dart
@@ -1,0 +1,15 @@
+part of 'accounts_bloc.dart';
+
+@freezed
+class AccountsEvent with _$AccountsEvent {
+  const AccountsEvent._();
+
+  const factory AccountsEvent.initialize() = Initialize;
+
+  const factory AccountsEvent.created({
+    required MyAccount account,
+    required AccountSource source,
+  }) = Created;
+
+  const factory AccountsEvent.loggedOut() = LoggedOut;
+}