+++ packages/espressocash_app/lib/features/app_lock/widgets/app_lock_menu_item.dart
@@ -5,7 +5,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import '../../../l10n/l10n.dart';
 import '../../../routes.gr.dart';
 import '../../../ui/profile_switch.dart';
-import '../src/bl/app_lock_bloc.dart';
+import '../services/app_lock_bloc.dart';
 
 class AppLockMenuItem extends StatelessWidget {
   const AppLockMenuItem({super.key});
@@ -13,8 +13,8 @@ class AppLockMenuItem extends StatelessWidget {
   @override
   Widget build(BuildContext context) => BlocBuilder<AppLockBloc, AppLockState>(
         builder: (context, state) => ProfileSwitch(
-          label: context.l10n.appLock,
-          subtitle: context.l10n.appLockDescription,
+          label: context.l10n.appLock_lblAutoLock,
+          subtitle: context.l10n.appLock_lblAutoLockDescription,
           value: state is AppLockStateEnabled,
           onChanged: (value) {
             final screen = value