+++ packages/espressocash_app/lib/features/balances/module.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:nested/nested.dart';
+
+import '../../di.dart';
+import 'services/balances_bloc.dart';
+
+class BalancesModule extends SingleChildStatelessWidget {
+  const BalancesModule({super.key, super.child});
+
+  @override
+  Widget buildWithChild(BuildContext context, Widget? child) => BlocProvider(
+        create: (context) => sl<BalancesBloc>(),
+        child: child,
+      );
+}