+++ packages/espressocash_app/lib/features/ramp/widgets/ramp_buttons.dart
@@ -5,9 +5,10 @@ import 'package:ramp_flutter/ramp_flutter.dart';
 
 import '../../../../../l10n/l10n.dart';
 import '../../../config.dart';
-import '../../../core/accounts/bl/account.dart';
-import '../../../core/balances/context_ext.dart';
+import '../../../l10n/device_locale.dart';
 import '../../../ui/button.dart';
+import '../../accounts/models/account.dart';
+import '../../balances/widgets/context_ext.dart';
 
 class AddCashButton extends StatelessWidget {
   const AddCashButton({
@@ -22,9 +23,10 @@ class AddCashButton extends StatelessWidget {
         child: CpButton(
           size: size,
           minWidth: 250,
-          text: context.l10n.addCash,
+          text: context.l10n.ramp_btnAddCash,
           onPressed: () {
             final configuration = _defaultConfiguration
+              ..selectedCountryCode = DeviceLocale.localeOf(context).countryCode
               ..defaultFlow = 'ONRAMP'
               ..userAddress =
                   context.read<MyAccount>().wallet.publicKey.toBase58();
@@ -52,7 +54,7 @@ class CashOutButton extends StatelessWidget {
         child: CpButton(
           size: size,
           minWidth: 250,
-          text: context.l10n.cashOut,
+          text: context.l10n.ramp_btnCashOut,
           onPressed: () {},
         ),
       );