+++ packages/espressocash_app/lib/main.dart
@@ -11,12 +11,12 @@ import 'package:provider/provider.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-import 'app/app.dart';
+import 'app.dart';
 import 'config.dart';
-import 'core/accounts/module.dart';
-import 'core/balances/module.dart';
 import 'core/dynamic_links_notifier.dart';
 import 'di.dart';
+import 'features/accounts/module.dart';
+import 'features/balances/module.dart';
 import 'logging.dart';
 import 'ui/splash_screen.dart';